requirements.txt
fastapi==0.111.0
langchain==0.1.17
langchain_community==0.0.36
python-dotenv==1.0.1
